pyproject.toml
[project]
name = "test-agent"
version = "0.1.0"
description = "Recipe agent testing with Scenario framework"
requires-python = ">=3.11"
dependencies = [
    "langwatch-scenario>=0.7.14",
    "openai>=2.8.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
]
